update activation layer logit

+++ storch_core/src/main/scala/torch/nn/modules/activation/LogSoftmax.scala
@@ -40,7 +40,7 @@ import torch.internal.NativeConverters.fromNative
   */
 final class LogSoftmax[D <: DType: Default](dim: Int) extends TensorModule[D]:
 
-  private val options = new LogSoftmaxOptions(dim)
+  private val options = new LogSoftmaxOptions(dim.toLong)
   options.dim().put(dim)
 
   override val nativeModule: LogSoftmaxImpl = LogSoftmaxImpl(options)
@@ -50,6 +50,7 @@ final class LogSoftmax[D <: DType: Default](dim: Int) extends TensorModule[D]:
   override def hasBias(): Boolean = false
 
   def apply(t: Tensor[D]): Tensor[D] = fromNative(nativeModule.forward(t.native))
+  
   def forward(input: Tensor[D]): Tensor[D] = fromNative(nativeModule.forward(input.native))
 
   override def toString = getClass().getSimpleName() + s"(dim=$dim)"
